// Final-molaris-main/project_molarisse-main/frontend/src/app/core/models/doctor-verification.model.ts
export interface DoctorVerification {
  id?: number;
  doctorId: number;
  status: 'pending' | 'approved' | 'rejected';
  address: string;
  cabinetAddress: string;
  yearsOfExperience: number;
  specialties: string[];
  postalCode: string;
  email: string;
  cabinetName: string;
  cabinetPhotoPath?: string;
  diplomaPhotoPath?: string;
  phoneNumber: string;
  message?: string;
  createdAt?: string;
  updatedAt?: string;
  // Profile picture properties
  firstName?: string;
  lastName?: string;
  hasProfilePicture?: boolean;
  hasProfilePictureChecked?: boolean;
} 